Bugfix of merge join operation (#158)
This branch address some problem with the keys used for the
merge. Depending on left_on/right_on and others, keys are
set NaN or to their other values if missing. The PR addresses
issues identified last week with the FRB code. It also solves all
bugs in the 69 tests for test_join.py