Groupby error checking (#154)
Closes #114 

For `df.groupby()`:
Error checking checks for required parameters(`by` only, since we don't support MultiIndex),  and types/validity of other parameters. In addition, it ensures `BodoError` is thrown with specs we currently do not support.

For `Groupby` functions:
For the built-in ones, check to make sure column datatypes are the ones we currently support.
For `agg()` and `aggregate()`, check to make sure `func` is supplied and specified columns are of valid types for the user defined func. Note, when multiple user defined functions are supplied for `agg()`, a column must be explicitly selected.

Some of these checks are based on current implementation constraints, so as we have more supports, some of the checks need to be removed.