Groupby: more general implementation and better C++ offloading
This makes the groupby code more general, so that for a given input
column, a list of functions can be applied to it, and different
functions can be applied to different columns. Using tuple notation
with groupby.agg(), a tuple of functions can be applied to one
column. Using dict notation, a dict specifying a function or list
of functions for each input column is supported. Any builtin
functions will now always run entirely in C++. For user-defined
functions, the update, combine, and eval portions run in generated
code and the rest in C++. Mixing builtins with user-defined
functions is now supported. See new tests in this PR for examples.

Mixing nunique and median with other functions is also supported.
What happens in this case is that the data is shuffled before
applying any function (because it is currently required for nunique
and median). So there is no combine step in this case.

cumsum, cumprod, etc. can now be used with groupby(...).agg(v)
where v is tuple or dict containing cumulative operations. Note
that semantically these cannot be mixed with regular aggregation
functions and a compile-time check has been addeed for that.

Closes #156
Closes #277
Closes #351
Closes #352