to_parquet: fix writing empty chunks in distributed mode (#978)
Previously, the index would not be written if it was size 0, which
happened on processes with empty chunks. This caused a mismatch
in the schema between parts that had the index and those that
didn't.

This patch fixes the issue by also writing an index if size is 0.
We pass a bool to C++ called `write_index` to tell it whether
the index has to be written or not.