Fix nightly tests (#1089)
- Nightly tests were not finding pytest.ini and were throwing
warnings about unrecognized pytest marks. Moved pytest.ini to
bodo directory (top-level directory of the bodo distribution package)
and modified MANIFEST.in and setup.py so that pytest.ini is
included in the Bodo package.

  Also modified azure-test-binary-linux-macos.yml to copy pytest.ini
  to the working directory where testing is triggered (otherwise
  pytest won't see it and there is no way to pass a pytest.ini location).

- Nightly tests were not finding runtests.py because it was not
included in the bodo package. Moved it to the bodo package
top-level directory so that it is included.