Optimize PQ file read (by creating file readers at time of use and parallelizing row counting) (#1504)
Earlier we used to create all the readers upfront and store them in a vector. This caused issues when there were a lot of files (and it would complain that too many files were open). Now we store the filenames up front in a vector instead and create the filereaders for them when they're needed (and get rid of them straight after).
Also the row counting process happened only on proc0. We've parallelized across all processes which results in a good speedup.