Runtime C++ IO error raise exceptions in python instead of crashing (#1469)
Implemented a sample C++ error handling approach for IO related operations (csv/json/pq dist/repl read/write).
The approach is this: throw std::runtime_errors with appropriate error messages whenever C++ code fails (while ensuring there are no hangs). Catch these exceptions/errors at the top-level C++ functions (that are called from python) in a try-catch clause. In case of error, use PySetErr_String to set the Python C API's error string to the error message from the exception/error. In python code, right after the call to a C++ function, call the new intrinsic check_and_propagate_cpp_exception, which checks if the error flag has been set (we did this when we called PySetErr_String in C++), and if it is, it raises an Exception in Python with the error message (which we'd set in C++ using PySetErr_String). We've also added status checks to a variety of Arrow and other functions (most importantly ValueOrDie() which crashes the program if called when the status is not ok), such that if the status is not ok, it throws a runtime error, which is caught at the top level C++ function and subsequently raised in Python. This solves an important problem that in cases like invalid credentials, unknown bucket region, etc., the arrow code would crash, which is bad in case of NBs (leads to an unrecoverable state). Now, it is caught before that can happen, and we can catch it and raise a readable error (in both C++ and importantly python), and avoid crashes.
The next aim should be to integrate this approach in all our codepaths for improved and unified error C++ handling throughout our code-base.