Require boost-cpp =1.74.0 (#1694)
We used to rely on arrow-cpp to require the right version of boost-cpp. Lately arrow-cpp has been inconsistent in listing boost-cpp versions in their dependencies/constraints. This caused an issue wherein the Bodo-binary was created using a version of arrow which required boost-cpp-1.72.0 and hence required this specific version of boost during execution. However, the arrow version installed on the platform didn't list constraints on the boost-cpp version, which is why it installed boost-cpp-1.74.0. Because of this, import bodo failed with:

from bodo.libs import hio 
ImportError: libboost_filesystem.so.1.72.0: cannot open shared object file: No such file or directory

This is a quick fix to address this issue.