Resolve read_csv 1 core issue (partially) (#1789)
Summary:
Still needs the objmode fix from numba, but this (sort of) takes care of the bottleneck from pandas.

Pandas' `read_csv` and Bodo's `read_csv` are not exactly equivalent, for instance in a column of `int64` if there is a missing entry, pandas would convert it to a `float64` column whereas Bodo would use a `Int64` type (nullable integers), etc. We discovered a performance bug with certain nullable types, notably `Int64`, in read_csv, i.e. when we specify the `Int64` dtype in `pd.read_csv`, the performance is very poor. Interestingly, if we do `pd.read_csv` without `dtype` argument and then simply do `df.astype` right after, we do not face the performance penalty. However, when reading strings, if we have missing entries, doing `df.astype` would convert those entries to literally the string values "nan". This is not desirable. Ideally we would use the nullable string type ("string") which would not have this issue, but unfortunately the performance is slow (in both `pd.read_csv` and `df.astype`). Therefore, we use a workaround here wherein we specify the `dtype` for strings (`str`) directly in `pd.read_csv` (there's no performance penalty, as per our testing), and specify the rest of the dtypes in the `df.astype` call.

Co-authored-by: Juan Galvez <juan@bodo.ai>