read_parquet support for columns with all nulls  (#1833)
Approach: Assign `pa.null()` to the `string_type`. During compilation we record information about which columns are this `pa.null()` type (in `null_col_map`). We add this information to the parquet IR node. Then in the` _gen_pq_reader_py` function we would have this information, and for these null columns, we would initialize them using `pre_alloc_string_array` (instead of using `ds_reader`) and set all the null bits for this array to 0 (i.e. specifying that they are null).