New Platform Build Pipeline (Weekly) and new channel for non-release binaries (#2303)
Created a new separate pipeline for building the platform binaries. This pipeline runs weekly (Sunday midnight) and on releases.
Created a new channel bodo.ai-platform-dev on our artifactory. All non-release binaries from this pipeline go to this new channel (hence not overwriting the binaries in bodo-binary channel as they do now), and the release binaries go to the bodo.ai-platform channel as before.
Also removed the suffix_step wherever it's not needed.