[BE-1147][BE-384] Refactor and simplify gb.nunique, and improve heuristic (#3139)
The heuristic now uses the estimated number of unique elements
if the nunique columns to decide whether to drop duplicates before
shuffling, and the global estimate of number of groups to decide
how to shuffle, whether on keys or keys+values. The latter is
necessary when the number of groups is small, to spread load.