[BE-3679] Sort `str_columns` in `determine_str_as_dict_columns` (and also before passing it in case of Iceberg) (#4356)
We discovered that the code generated on different ranks was different because the order of the columns on different ranks was different (since we were using a set). This PR sorts the list of string columns before passing it to determine_str_as_dict_columns (in case of Iceberg). As a precaution, we also sort the list inside the function. Parquet seems to pass in a list in deterministic order, so no change in required in the read-parquet infrastructure.

* Sort str_columns in determine_str_as_dict_columns (and also before passing it in case of Iceberg)

* Add test to protect against future regressions