[BE-3538] Add a `pytest` fixture to assume-role and update AWS credentials in e2e tests to fix issues with S3 Reads/Writes (#4372)
Our Codebuild project does have a service role that has access to the required S3 buckets, etc. The way this usually works is that during S3 read, for instance, Arrow (the AWS SDK under the hood) goes to the instance metadata service and gets the required token to make the S3 API calls. However, Codebuild seems to have some throttling at the token generation level when all ranks request them, leading some ranks to end up without a valid token. To circumvent this, we retrieve a set of temporary AWS credentials once before every test, so that all ranks can read them directly from the environment and don't run into issues due to throttling. In our tests we usually call mpiexec in a subprocess, which should inherit these environment variables. The reason this needed to be a pytest fixture is that if we were to assume the role in run_nightly_tests.sh once, it only allows us to get credentials that are valid for 1hr (limitations of role chaining: https://aws.amazon.com/premiumsupport/knowledge-center/iam-role-chaining-limit/). We can get the token directly from ECS Metadata service (like https://stroobants.dev/use-metadata-iam-role-in-your-docker-build.html suggests) which are supposed to be valid for 6hrs by default. However, in case of Codebuild, it seems like they're only valid for 1hr. See more experiments and details here: https://bodo.atlassian.net/browse/BE-3538.

We’ve made it so that Bodo developers (part of codebuild-perms user group on AWS), have ability to assume the BodoEngineNightlyRole role, and so does the Codebuild Service role, so it should be pretty seamless.