[BE-4219] Fix Iceberg-Related Errors Discovered During Insight Testing (#4759)
This PR fixes some Iceberg-related issues that we noticed while testing some  queries. These include
- Enabling and setting S3FileIO as default on all Catalogs. 
We noticed that newer versions of Spark, such as EMR's Spark + Iceberg, seems to automatically use S3FileIO. This FileIO will encode S3 paths as `s3://` in Iceberg metadata files, which HadoopFileIO does not support. To ensure that our Iceberg connector is generally compatible and can read such tables, it is better if we use S3FileIO by default. Note that S3FileIO can read `s3a://` and `s3n://` paths, so it is compatible with HadoopFileIO written tables, just not the opposite.

- Warn when the warehouse parameter is not given
We noticed some odd behavior when the warehouse parameter is not provided during read, especially for Nessie catalog. While it is not always necessary, we should at least notify that the parameter should be included.

- Printing Java Exceptions by Default
It seems like we no longer print the entire stacktrace, only the top-level error, in the read operations. They seems to contain a lot of useful information, especially for generic cases like RuntimeIOError. Thus, I think it is better to include the Java exception in the IcebergError than hiding when Numba Developer mode is not enabled.

- Being able to read when the table is empty (just created)
There are a couple of assumptions that need to be made to get around it. Since there is no snapshot, we need to return a default value. Furthermore, the parquet dataset is null so we need to fix the dictionary encoding detection.

- Other various QOL improvements