[BE-4220] Add useful scripts for query testing; Update build from source on platform steps (#4768)
(As title)
With these instructions, the existing scripts for building from source on the platform shouldn't be required, so I've removed them to avoid any confusion.
This PR also adds a couple of scripts that will be useful for testing the Snowflake queries:
(i) `BodoSQLWrapper.py`, which is a script that can take a `.sql` files and execute them.
(ii) `snowflake_utils.py`, which has some basic utility functions for exploring table statistics, executing the query in Snowflake, etc. 
We also have 2 notebooks with examples of how to use these scripts.