[BE-4287] Enable overflow during cast from `Date` to `Timestamp` in Snowflake Read (temporary workaround) (#4795)
(As title)

We have limitations around the “Date” type (in BodoSQL particularly since we don’t have a Date type) because of which we usually convert it to `datetime64[ns]` everywhere. This used to generate an astype call which restricts limit pushdown. To work around this, we added a flag called `_bodo_read_date_as_dt64` to `read_sql` (see this PR: https://github.com/Bodo-inc/Bodo/pull/4776), which pushes the cast from `date32[day]` to `datetime64[ns]` to the ArrowDataFrameReader. I know this wasn’t designed or tested for this use case, but it usually works fine. There’s an issue that some dates cannot in fact be represented using `datetime64[ns]` (max is `2262-04-11 23:47:16.854775807`). When Arrow sees a date beyond this, the cast function rightfully complains and errors out. This cast seems to “work” when using Bodo's `astype` (that uses the `pd.Timestamp` constructor), i.e. it fills it with garbage data but doesn’t error out.
happens to have dates like `9999-02-01` in their tables (they probably fill it with nonsense values when they don’t know what the date is, or to indicate nulls, etc.). This is technically a valid date as far as the date type is concerned, so we can’t really tell that they should change their data.
Arrow cast can be configured to allow overflow, which doesn’t error out and just fills the buffer with garbage values. This essentially matches the behavior of the astype conversion we were doing earlier.

I should note that Pandas' `astype` also raises an `OutOfBoundsDatetime` exception. Bodo's `astype` is the one that's flawed and doesn't validate the input. This can technically cause dangerous correctness issues. Therefore, we have a runtime environment variable `BODO_DISABLE_UNSAFE_DATE_TO_TS_CAST`, which when set to (any value), will disable this unsafe cast behavior.

This PR should definitely be reverted once we have a proper Date type 😅.