[BE-4270] Fix M1 Conda Builds and Add Debugging Parameter (#4800)
Fixes M1 Conda Builds to actually build a M1 package instead of a x86 Mac package, by adding an additional flag to indicate M1 builds to the build script.

In addition, this PR adds some additional variables to enable Mac, M1, and Pip builds when triggering CI manually, so that it is easier to test build changes in the future with the Nightly CI pipeline. Before this PR, Nightly CI would always skip Mac Conda and all Pip builds, making it hard to test the build step until we actually performed a release.