[BE-4343] Simple base classes for representing parts of the generated code (#4847)
This should help us to build a base of abstractions around the code
generation so there are fewer raw strings being passed around with no
context. The various raw/code attributes are included as a way to just
escape the entire system and insert raw code which is what existing code
would end up doing at the current moment.

These classes are written in Kotlin. Kotlin is a JVM based language
transpiler that's like a TypeScript for Java code. It allows writing
Java code faster while having no impact on performance,
interoperability, or readability. We can rewrite these in Java and use
them from Java code, but Kotlin will likely increase development
productivity and is a generally used tool in the industry for Java
development.

[run CI]