[BE-4331] Drop NAs from dictionary of dict-encoded arrays during deduplication step; Ensure deduped dictionary during Parquet/Iceberg/Snowflake write (#4849)
(As title)
The basic problem we're trying to solve is that `Arrow::Parquet` doesn’t support writing a dictionary encoded array with nulls in the dictionary ([see this](https://github.com/apache/arrow/blob/ea6875fd2a3ac66547a9a33c5506da94f3ff07f2/cpp/src/parquet/arrow/path_internal.cc#L753)).
While we don’t usually store nulls in the dictionary, it might still happen in one of the BodoSQL kernels.

To fix this, we've modified `update_local_dictionary_remove_duplicates` to drop NAs in the dictionary by default (since that is _always_ the desired behavior). Therefore, now `has_deduped_local_dictionary` additionally means that there aren't any nulls in the dictionary. This takes care of a way to drop nulls from the dictionary.

Now, for the actual write functions, we need to make sure that all dictionary encoded arrays have `has_deduped_local_dictionary`, and if not, then call `drop_duplicates_local_dictionary` on them to make it so that they do. We could've done it in the `pq_write` function, however, this function is used as a helper function in a lot of cases, e.g. partitioned parquet write, iceberg write, snowflake write, etc. The array ownership can get a bit messy, and therefore we opted to not modify the dictionary in `pq_write` and instead do it during the codegen where the ownership is clearer. Also, this means that we have to do it only _once_ for each array, instead of doing it multiple times (as would've been the case for partitioned parquet write or snowflake write, where we write a subset of the table at a time). This has been implemented for Parquet (both regular and partitioned) write, Iceberg write and Snowflake write.