[BE-4386] Fix up bodosql code generation when append is used (#4902)
The append method was used to be compatible with old code, but it still
assumed that either the append would be done in a single line or with a
chain of append calls. The aggregate generation code uses chained
appends with multiple append calls so it was erroneously believing that
the appends were for different statements rather than the same one.

The append call now just appends to the last operation if that operation
is raw code or generates a new raw code operation if something else was
used. The raw code operation then handles multiple lines with multiple
statements and indentation properly.

[run CI]