Add tpch tests to the java code for testing code generation (#4890)
This copies the tpch tests from the test plans and includes expected
outputs for the unoptimized plan, the optimized plan, and the end code
generation.

The purpose of this is to help track changes closer to where the change
happens and offer a way to visually compare the differences. The python
tests are important for testing that the generated code actually works
and these tests don't test the functionality of the code, just that the
code looks like how we expect it to.