Remove getName() method from RexNodeVisitorInfo (#4932)
This method has been marked as not used. I've gone through each of the
places where it was used and removed its usage. While the method itself
wasn't used in any meaningful way for code generation, it was
recursively passed around so there was no way to identify when it was
used and when it wasn't.

There were two places in logical projection where it was used
transitively. It was defined and then immediately utilized. These have
been separated to be more explicit and we can refactor them later.
There's no reason to have the extra attribute just for these two
locations inside of the logical projection code.