[BE-4400] Keep outer sides nulls locally (and drop when on the inner side) in case of SQL joins (#4931)
(As title)
In case of SQL joins, nulls are not considered equal to each other. In cases where there are a series of outer joins, a large amount of nulls are usually generated even though they cannot match with any other rows. Originally, we sent all nulls to the same rank, which can create huge skew in the amount of data each rank has to process. We saw this very recently with 's Query 3. To handle nulls better, we have made the following changes to our join/shuffle infrastructure:
- Added a new function that can essentially do a bitwise AND of multiple null bit masks. For types that don’t already have a bit mask (e.g. NUMPY), we create a temporary one. I’m assuming NAs would need to be considered nulls in this case. The purpose of this is to check if any of the keys in a row is null, because if so, that row cannot match with anything in the `!is_na_equal case` and we can use that information to make smarter shuffle decisions.
- In `equi_join_shuffle`, if `!is_na_equal`, we create the null-bit mask for the keys using the new function described above. We have modified `coherent_shuffle_table` and `mpi_comm_info::set_counts` to use the bit mask to drop the nulls altogether if the table is on the inner side of a join, and keep the nulls on the same rank in case it is on the outer side of a join.

See issue description for more details.