[BE-4279] Switch RelationalAlgebraGenerator to use withExpand = false (#4943)
This switches the RelationalAlgebraGenerator to use withExpand = false
and to then use the `SubQueryRemoveRule` to remove subqueries and the
`RelDecorrelator` to remove the correlation nodes.

It also adds a new planner rule that will remove RexOver nodes from join
conditions. The `SubQueryRemoveRule` has a tendency to put these into
join conditions when the previous code would not. Since we can't handle
these nodes inside of join conditions, this planner rule will move it to
a projection before the join and then remove the field after the join is
complete.

This updates the existing set of tests to work. The ultimate goal of
this is to enable `ANY/ALL` behavior in a subsequent PR. The other
reason is because the Calcite developers have made clean that withExpand
= false is intended to be the default behavior and will be the option
that's generally supported in the future.

In addition, this commit removes unused test dependencies that
interfered with the execution of the test framework through maven. The
tests could be run through IntelliJ, but Maven didn't seem to pick them
up because of the other test frameworks that existed but also had no
tests defined for them.

This also extends the test framework for planner rules and code
generation to include a way to update the tests. This can be done with:

    UPDATE_EXPECT=1 mvn test -fae

This sets an environment variable in the test framework that will write
the actual output to the expected output files. `-fae` tells the test
framework to continue even when a failure is encountered. The tests
still get marked as failed because maven seems to cache successes so
succeeding causes the tests not to run again when the environment
variable is gone.