Move cache for dict-encoded like kernel to C++ (#4992)
(As title)
In the BodoSQL like kernel, when both the array and the pattern are dictionary encoded arrays, we cache the outputs for combination of indices for the two arrays. We used to use a regular dictionary (numba), which is not always the most performant. This PR moves the cache to a C++ unordered map. This gives a ~10-30% performance Query 2, and should be generally applicable.