[BE-4279] Support most ANY/ALL comparisons with subqueries (#4945)
This adds support for most ANY/ALL comparisons with subqueries. The
ANY/ALL comparisons get translated into various join conditions based on
the `SubQueryRemoveRule` and `RelDecorrelator`.

This supports most functionality for subqueries, but support for the
`LIKE` operator isn't part of the existing Calcite functionality. A test
has been included for that and it has been commented out so we can get
it working later.