[BE-4454] Merge the calcite-core package with calcite-application (#4988)
In order to simplify the build and testing, I'm merging these two
projects. They aren't deployed separate from each other and the current
setup just increases the chance of confusion about where to look for a
certain file.

If there's a compelling reason for us to split these in the future for
deployment reasons, we can do that then.