[BE-4358] Remove columnNamesStack from the codegen visitor (#5008)
This removes `columnNamesStack` from the codegen visitor. This was used
to retrieve the list of column names for an input node, but it had a
cost to maintain and the information was readily available on each
`RelNode` through `getRowType()`.

Since this variable isn't needed and just adds complexity to the code,
it's better to just remove it.