[BE-4169] Port bodo sql parser to our internal calcite project (#5035)
This ports the separated bodo sql parser to our own internal calcite
project so it can be modified as part of this repository instead of in a
fork.

This is related to an effort to upgrade the calcite version to the
latest. The easiest way to do this is to just stop having modifications
in the fork and confine the parts that are separate into this project
where they are actually used. The parser was already fairly segmented so
it was easier to pull this.

Of note, this process is not complete. As part of this commit, there are
a few files copied from the core package of calcite. The header files
have been maintained on these files so they should fulfill the
requirements of the license. Overall, the eventual goal is to completely
remove these files and modifications to the core parser and instead use
the one included in the upstream calcite's core package. This will
require further modifications to maven, but those changes aren't part of
this because there's no point since we've modified those files anyway.

This also means that, if we upgrade, we will not gain any benefit from
the parser code because we've wholesale copied the parser template into
our own project. But, this makes it easier to diff the upstream file and
our own modifications and recreate those modifications in our own
version and test them in our own tests.

Since the parser is the only thing the bodo module was used for in the
fork, this also removes a dependency on that fork.