[BE-4486] Move various rex visitor implementations into their own class (#5039)
This moves `visitRexNode` and the various helper classes into their own
class. In the Calcite code, these are generally called
`RexToXXXTranslator` classes which translate a `RexNode` into the native
format.

This has a few benefits which aren't immediately obvious from looking at
this commit.

First, this allows us to more easily split the implementations of the
`RelNode` classes into their own classes. Right now, all classes emit
their code from the `PandasCodeGenVisitor`. We would like to be able to
organize the generative code with the relational node itself using
`PandasRel` as the unifying area of the code. Doing this allows us to
move towards refactoring the `RelNode` values too.

The benefit of refactoring the `RelNode` values is that we get to start
utilizing what Calcite calls conventions. Conventions are areas of the
plan that are marked for a specific calling convention. This is how
Calcite handles different domains and converting data between those
domains. In our own world, we would have the Snowflake convention and
the Pandas convention.

When we have these conventions, we can then more easily utilize the
`VolcanoPlanner` to determine which push downs would be optimal and to
automatically perform those optimizations as part of its normal duties.
It will then be easier to utilize more of the existing planner rules and
those planner rules will automatically determine various push downs
rather than just running the `HepPlanner` multiple times and hoping for
a push down optimization.

Once we have an in-use `VolcanoPlanner`, this opens up the entire world
of relational expression optimization.

The current implementation of this is very hacky. The purpose of the
translator is to hold the current `RelNode` and the input to the
`RelNode` for it to generate code. There's a lot of code that needs to
be refactored within it and some sections of this are just to allow the
code to be moved so those sections can then be isolated and refactored
later.

[run CI]