[BE-4169] Separate planner details into a separate class (#5062)
This change separates the planner components of the
`RelationalAlgebraGenerator` into its own reusable class. The class has
been integrated with the test suite so they share the common code.

This will allow us to isolate changes to the planner into a single file
so that changes to it are less confusing. In particular, we will want to
make custom modifications to the upstream planner so that we can inject
our own versions of the components such as the validator into the
pipeline. This change moves the current setup into the class that will
need to be created and also allows this to be reused by other
components such as the test suite.

As part of this change, there are some modifications to the test file
outputs. This is because some drift accumulated between the two pieces
of duplicated code so the tests were not accurately rendering the
current set of planner rules. Since this change consolidates them, the
tests changed to their correct outputs given the planner rules.

Another small addition is the introduction of using the `programs`
feature of the planner. Programs are certain transformations we define
and are also sometimes referred to as "phases" or "stages". The two
phases are the subquery removal phase and the optimizer phase.