[BE-4169] Copy over the planner so it can be modified directly (#5077)
This copies over the planner and moves it into the same folder structure
as the calcite project. The planner is the primary entrypoint to calcite
and controls the various components that will exist throughout the query
lifecycle. This includes the creation of the components.

While a new parser creation can be injected through the framework
config, most components can only be modified through pre-existing
options and aren't as expansive as we would want.

This copies over the planner implementation with modifications to use
Bodo components and defaults. This will allow us to inject our own
versions of each of the components and control the lifecycle of the
query more thoroughly. We'll still be able to utilize most upstream
changes, but this lets us control how they are used and initialized.

This change is completely compatible with upstream calcite 1.30.