[BE-4486] Wrap supported logical operations in Pandas version of them (#5104)
This wraps the logical operations supported by the code generator into
Pandas versions. These new operations use the pandas convention in
preparation of switching to the VolcanoPlanner.

The convention doesn't do anything except for just exist and the
converters aren't necessarily correct implementations, but the correct
implementations aren't compatible with the `HepPlanner`.

An extra program has been added to the end that converts each of the
logical operations to the pandas operations so it doesn't conflict with
existing planner rules. The new operations, while they have the
capability to emit themselves, presently do not and rely on being
intercepted by the code generator. This was to reduce the amount of code
moving while still enabling that refactor in the future.

One aspect of this change that I'm not quite happy with but will likely
involve a more involved refactor in the future are the operations
related to table modification, creation, and the target table scan.
These operations are direct physical operations on an external source
and will likely require some refactoring to decide how these should
look. This isn't the change for updating those and so they're mostly
just copied to preserve the current functionality.