[BE-4531] Create Snowflake convention and SnowflakeToPandasConverter (#5153)
This creates the Snowflake convention and a SnowflakeToPandasConverter
that get registered to transform the query into its final destination.
The plan now properly conveys that boundaries between different systems
so the planner is aware of those conventions and we can begin to perform
optimizations around them.

The rules to convert snowflake operations to the pandas convention have
been removed in favor of using the snowflake convention. Presently, the
`SnowflakeTableScan` doesn't include the convention by default because
it interferes with the `HepPlanner`. We use another program to add the
convention before running the `VolcanoPlanner`.

The code generation for the aggregate pushdown has also been moved into
the converter and the preserve case hack has been removed. At the same
time, the nodes still lie some amount about their types. The table
exposes the lowercase names so the validator gets the casing right on
things like wildcards. The actual snowflake nodes expose the correct
column name. The converter then marks that the columns become lowercase
to signal the same behavior that `pd.read_sql` does (since that's what
that node is). When we add the converter, we also add a `PandasProject`
to restore the names to the proper casing as expected by the plans. You
may find that the generated pandas code sometimes reads a table just to
rename all of the columns to uppercase variants. This is part of that
code path.

A lot of that behavior will disappear once all planner rules are in the
volcano planner as the generated projections will end up being smarter.