[BE-4548] Separate `MemInfo` and data allocations (and minor refactor of `_meminfo.h` for readability) (#5170)
(As title)
This is a pre-requisite change to the integration of MemInfo allocation (NRT infrastructure) and the Buffer Pool (#5143). Moving forward the "data" part of the allocation will be moved to the Buffer Pool, while the MemInfo allocation will stay as it is now (for the most part).
While doing this, I also refactored the _meminfo.h to (hopefully) make it more readable. In particular, I've added docstrings, moved functions closer to their usage and added boundaries for different functionalities.