[BE-4534] Add VolcanoPlanner as an option and include output in tests (#5174)
This change introduces a new program for using the volcano planner as
the sole optimizer. This is only included for unit tests and outputs its
own file for testing purposes.

In addition to the volcano planner, a custom cost structure has been
included. The custom cost structure includes values for cpu, io, and
memory and compares values based on a calculation utilizing these
values. The output for both optimized versions of the plan now uses the
new cost structure for comparison purposes.

As an initial test of using the metadata providers, a custom provider
has been added for pandas nodes that inserts the number of ranks as part
of the parallelism attribute. This allows nodes with `PandasRel` to know
how many ranks are available for the purposes of cost generation. In
addition, the ability to override this value is included as part of the
`PandasRel` interface. The actual number of ranks isn't passed through
as it just defaults to 2 at the current moment.

The output for the optimized plans has been changed to both include
costs and include ids when there are duplicate nodes with the same id.
The ids have been normalized to reduce churn from changing plans. Only
ids for nodes with duplicates are included because we don't want a
change in the plan in one place to cause all of the ids to change and
make the diff unreadable. Similarly, costs are also per operation rather
than Calcite's default of a cumulative operation. The cumulative
operation means that every operation above one operation would change if
we modified the cost structure even if the overall plan didn't change.
This change prevents that from changing the diff.

No other relational operators have been modified to give a custom cost.
That will be a future issue and project.