Moving certain long running tests to the nightly and removing duplicates (#5212)
* Removes the TPCH plans from the test plans

These tests are included in the Java tests and these ones are also
unreasonably slow because they read parquet files to determine the
schema rather than having the actual schema available.

The Java tests are much faster, contain auto-update functionality, and
also are generally closer to the code that's being modified when they
change.

* Mark the filter or rules test as slow since it is slow

This test and all of the tests in this directory are pretty slow. These
tests seem to go through the entire process of parsing, validating,
planning, and executing queries against both Bodo and Spark for the
purpose of learning whether optimization rules triggered or not.

These should really be Java unit tests as they'll run much faster there.
The rest of this package is slow tests though so marking it as slow for
now to aid CI speed.

* Disables the window function tests as part of the normal CI

These tests are absurdly long to the point where they have to modify the
default timeout. They don't really belong in the normal CI if they're
that long to do and should instead be part of a consideration for why
the compiler takes so long with these particular patterns.

The specific attributes these tests are looking for are also likely
better as Java tests. Revisit potentially just removing these tests and
putting them into Java as we're modifying that section of code.